hotfix(#68)[several bugs related to typescript and eslint]
