+++ client/src/components/Organisms/Global/Table/Table.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../bemit/settings/_colors.scss';
+@import '../../../../bemit/settings/colors';
 
 .Table {
   $block: &;
